refactor home cell view
